+++ PocketCartApp.Domain/Domain Models/Receipt.cs
@@ -19,5 +19,6 @@ namespace PocketCartApp.Domain.Domain_Models
         public double total { get; set; }
         public DateTime PaidAt { get; set; }
         public string? currency {  get; set; } //WILL BE A DROP DOWN MENU N SEARCH FOR THEM TO CHOOSE A CURRENCY, ALLOWS FLEXIBILITY
+        public string? PdfPath { get; set; }
     }
 }