+++ PocketCartApp.Service/Implementation/ShoppingCartService.cs
@@ -13,6 +13,7 @@ using System.Reflection.Metadata;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Routing.Constraints;
 
 namespace PocketCartApp.Service.Implementation
 {
@@ -74,7 +75,7 @@ namespace PocketCartApp.Service.Implementation
             {
                 return new ShoppingCartDTO
                 {
-                    Products = new List<ProductInShoppingCart>(),
+                    ProductsInCart = new List<ProductInShoppingCart>(),
                     TotalPrice = 0
                 };
             }
@@ -90,14 +91,50 @@ namespace PocketCartApp.Service.Implementation
 
             ShoppingCartDTO model = new ShoppingCartDTO
             {
-                Products = allProducts.ToList(),
+                ProductsInCart = allProducts.ToList(),
                 TotalPrice = totalPrice
             };
 
             return model;
         }
 
-        public bool PrintReceipt(string userId)
+
+        public ShoppingCart? GetByUserIdWithIncludedProducts(string userId)
+        {
+            return _shoppingCartRepository.Get(
+                selector: x => x,
+                predicate: x => x.CashierOnShift == userId,
+                include: x => x
+                    .Include(z => z.ProductsInCart)
+                    .ThenInclude(p => p.Product)
+            );
+        }
+
+        public void UpdateQuantity(string userId, Guid productId, double quantity)
+        {
+            var cart = GetByUserIdWithIncludedProducts(userId);
+
+            if (cart == null)
+                throw new Exception("Shopping cart not found.");
+
+            var item = cart.ProductsInCart
+                .FirstOrDefault(x => x.ProductId == productId);
+
+            if (item == null)
+                throw new Exception("Product not found in cart.");
+
+            if (quantity <= 0)
+            {
+                cart.ProductsInCart.Remove(item);
+            }
+            else
+            {
+                item.quantity = quantity;
+            }
+
+            _shoppingCartRepository.Update(cart);
+        }
+        public Guid PrintReceipt(string userId)
         {
             var userCart = _shoppingCartRepository.Get(selector: x => x,
                                              predicate: x => x.CashierOnShift == userId,
@@ -107,7 +144,7 @@ namespace PocketCartApp.Service.Implementation
                 userCart.ProductsInCart == null ||
                 !userCart.ProductsInCart.Any())
             {
-                return false;
+                return Guid.Empty;
             }
 
             double totalPrice = 0;
@@ -132,16 +169,20 @@ namespace PocketCartApp.Service.Implementation
             _receiptRepository.Insert(receipt);
 
             // PDF GENERATION
-            GenerateReceiptPdf(receipt, userCart);
+            var pdfPath = GenerateReceiptPdf(receipt, userCart);
+            receipt.PdfPath = pdfPath;
 
-            // CLEAR CART
-            userCart.ProductsInCart.Clear();
-            _shoppingCartRepository.Update(userCart);
+            _receiptRepository.Update(receipt);
 
-            return true;
+            foreach (var item in userCart.ProductsInCart.ToList())
+            {
+                _productInShoppingCartRepository.Delete(item);
+            }
+
+            return receipt.Id;
         }
 
-        private void GenerateReceiptPdf(Receipt receipt, ShoppingCart shoppingCart)
+        private string GenerateReceiptPdf(Receipt receipt, ShoppingCart shoppingCart)
         {
            string folderPath = Path.Combine(
                  Directory.GetCurrentDirectory(),
@@ -153,6 +194,8 @@ namespace PocketCartApp.Service.Implementation
                 {
                     Directory.CreateDirectory(folderPath);
                 }
+
+                string fileName = $"Receipt-{receipt.Id}.pdf";
 
                 string filePath = Path.Combine(
                     folderPath,
@@ -170,7 +213,6 @@ namespace PocketCartApp.Service.Implementation
 
                 document.Add(new Paragraph("RECEIPT"));
                 document.Add(new Paragraph("-------------------"));
-                document.Add(new Paragraph($"Receipt ID: {receipt.Id}"));
                 document.Add(new Paragraph($"Currency: {receipt.currency}"));
                 document.Add(new Paragraph($"Date: {DateTime.Now}"));
                 document.Add(new Paragraph(" "));
@@ -198,6 +240,8 @@ namespace PocketCartApp.Service.Implementation
                 );
 
                 document.Close();
+
+            return "/receipts/" + fileName;
             }
 
         public ShoppingCart? GetById(Guid id)
@@ -205,6 +249,12 @@ namespace PocketCartApp.Service.Implementation
             return _shoppingCartRepository.Get(selector: x => x,
                                                        predicate: x => x.Id.Equals(id));
         }
+
+        public ShoppingCart Update(ShoppingCart shoppingCart)
+        {
+            return _shoppingCartRepository.Update(shoppingCart);
+        }
+
 
         public void ClearCart(string userId)
         {
@@ -215,6 +265,7 @@ namespace PocketCartApp.Service.Implementation
             {
                 throw new Exception("Shopping cart not found for the user.");
             }
+            
 
             var cartItems = _productInShoppingCartRepository.GetAll(selector: x => x,
                                                         predicate: x => x.ShoppingCartId.Equals(shoppingCart.Id)).ToList();